Development

+++ iStudy-services/discipline-ms/src/main/java/io/github/abeatrizsc/discipline_ms/services/DisciplineService.java
@@ -3,39 +3,55 @@ package io.github.abeatrizsc.discipline_ms.services;
 import io.github.abeatrizsc.discipline_ms.domain.Discipline;
 import io.github.abeatrizsc.discipline_ms.dtos.DisciplineRequestDto;
 import io.github.abeatrizsc.discipline_ms.enums.DisciplineCategoryEnum;
-import io.github.abeatrizsc.discipline_ms.exceptions.DisciplineNameConflictException;
-import io.github.abeatrizsc.discipline_ms.exceptions.DisciplineNotFoundException;
+import io.github.abeatrizsc.discipline_ms.exceptions.NameConflictException;
 import io.github.abeatrizsc.discipline_ms.mapper.DisciplineMapper;
 import io.github.abeatrizsc.discipline_ms.repositories.DisciplineRepository;
+import io.github.abeatrizsc.discipline_ms.utils.AuthRequestUtils;
+import jakarta.ws.rs.NotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
 @AllArgsConstructor
 public class DisciplineService {
     private DisciplineRepository repository;
+    private AuthRequestUtils authRequestUtils;
 
     @Transactional
-    public void save(DisciplineRequestDto requestDto) throws DisciplineNameConflictException {
+    public void save(DisciplineRequestDto requestDto) throws NameConflictException {
         Discipline discipline = DisciplineMapper.INSTANCE.convertDtoToEntity(requestDto);
 
-        if (repository.findByName(discipline.getName()).isPresent()) {
-            throw new DisciplineNameConflictException();
+        String userCreator = authRequestUtils.getRequestUserId();
+
+        discipline.setCreatedBy(userCreator);
+
+        if (disciplineNameAlreadyExists(discipline.getName(), discipline.getCreatedBy())) {
+            throw new NameConflictException("subject");
         }
 
         repository.save(discipline);
     }
 
     @Transactional
-    public Discipline update(String id, DisciplineRequestDto requestDto) {
-        Discipline discipline = repository.findById(id).orElseThrow(DisciplineNotFoundException::new);
+    public Discipline update(String id, DisciplineRequestDto requestDto) throws NameConflictException {
+        Discipline discipline = findById(id);
+
+        if (!authRequestUtils.isRequestFromCreator(discipline.getCreatedBy())) {
+            throw new SecurityException();
+        }
+
+        if (disciplineNameAlreadyExists(requestDto.getName(), discipline.getCreatedBy())) {
+            throw new NameConflictException("subject");
+        }
 
         discipline.setName(requestDto.getName());
         discipline.setCategory(requestDto.getCategory());
+        discipline.setIsCompleted(requestDto.getIsCompleted());
 
         repository.save(discipline);
 
@@ -44,24 +60,62 @@ public class DisciplineService {
 
     @Transactional
     public void delete(String id) {
-        Discipline discipline = repository.findById(id).orElseThrow();
+        Discipline discipline = findById(id);
+
+        if (!authRequestUtils.isRequestFromCreator(discipline.getCreatedBy())) {
+            throw new SecurityException();
+        }
 
         repository.delete(discipline);
     }
 
     public List<Discipline> findAll() {
-        return repository.findAll();
+        return repository
+                .findAll()
+                .stream()
+                .filter(d -> Objects.equals(d.getCreatedBy(), authRequestUtils.getRequestUserId()))
+                .toList();
     }
 
     public Discipline findById(String id) {
-        return repository.findById(id).orElseThrow(DisciplineNotFoundException::new);
+        Discipline discipline = repository.findById(id).orElseThrow(() -> new NotFoundException("Subject"));
+
+        if (!authRequestUtils.isRequestFromCreator(discipline.getCreatedBy())) {
+            throw new SecurityException();
+        }
+
+        return discipline;
     }
 
-    public Optional<List<Discipline>> findByCategory(DisciplineCategoryEnum category) {
-        return repository.findByCategory(category);
+    public List<Discipline> findAllByCategory(DisciplineCategoryEnum category) {
+        return repository
+                .findAllByCategory(category)
+                .stream()
+                .filter(d -> Objects.equals(d.getCreatedBy(), authRequestUtils.getRequestUserId()))
+                .toList();
     }
 
-    public Optional<List<Discipline>> findByNameLike(String query) {
-        return repository.findByNameLike(query);
+    public Optional<Discipline> findByNameAndCreatedBy(String name, String userId) {
+        return repository.findByNameAndCreatedBy(name, userId);
+    }
+
+    public List<Discipline> findByNameContaining(String query) {
+        return repository
+                .findByNameContaining(query)
+                .stream()
+                .filter(d -> Objects.equals(d.getCreatedBy(), authRequestUtils.getRequestUserId()))
+                .toList();
+    }
+
+    public List<Discipline> findByIsCompletedTrue() {
+        return repository
+                .findByIsCompletedTrue()
+                .stream()
+                .filter(d -> Objects.equals(d.getCreatedBy(), authRequestUtils.getRequestUserId()))
+                .toList();
+    }
+
+    public Boolean disciplineNameAlreadyExists(String newName, String userId) {
+        return repository.findByNameAndCreatedBy(newName, userId).isPresent();
     }
 }
